+++ src/EZ-Template/drive/drive.cpp
@@ -230,24 +230,37 @@ void Drive::reset_drive_sensor() {
   }
 }
 
-int Drive::right_sensor() {
+double Drive::right_sensor() {
   if (is_tracker == DRIVE_ADI_ENCODER)
     return right_tracker.get_value();
   else if (is_tracker == DRIVE_ROTATION)
     return right_rotation.get_position();
+  // double position=0;
+  // for_each(right_motors.begin(),right_motors.end(),[&position](pros::Motor m){
+  //   position+=m.get_position();});
+  // position/=right_motors.size();
+  // return position;
   return right_motors.front().get_position();
 }
+
 int Drive::right_velocity() { return right_motors.front().get_actual_velocity(); }
 double Drive::right_mA() { return right_motors.front().get_current_draw(); }
 bool Drive::right_over_current() { return right_motors.front().is_over_current(); }
 
-int Drive::left_sensor() {
+double Drive::left_sensor() {
   if (is_tracker == DRIVE_ADI_ENCODER)
     return left_tracker.get_value();
   else if (is_tracker == DRIVE_ROTATION)
     return left_rotation.get_position();
+  // double position=0;
+  // for_each(left_motors.begin(),left_motors.end(),[&position](pros::Motor m){
+  //   position+=m.get_position();});
+  // position/=left_motors.size();
+  // std::cout<<"left position: "<<position<<"left front position: "<<left_motors.front().get_position()<<std::endl;
+  // return position;
   return left_motors.front().get_position();
 }
+
 int Drive::left_velocity() { return left_motors.front().get_actual_velocity(); }
 double Drive::left_mA() { return left_motors.front().get_current_draw(); }
 bool Drive::left_over_current() { return left_motors.front().is_over_current(); }
